plugin structure refactor
