+++ examples/dataprompt.config.js
@@ -1,7 +1,14 @@
 /*
   This file is not needed if you only want the default values.
   export a DatapromptConfig object
+
+  import type { DatapromptConfig } from 'dataprompt'
 */
+import { fsPlugin } from 'dataprompt/plugins/fs'
+
 export default {
-  promptsDir: 'prompts'
+  promptsDir: 'prompts',
+  plugins: [
+    fsPlugin()
+  ]
 }
