+++ package.json
@@ -11,9 +11,21 @@
       "import": "./dist/index.js",
       "types": "./dist/index.d.ts"
     },
-    "./firebase": {
-      "import": "./dist/firebase/index.js",
-      "types": "./dist/firebase/index.d.ts"
+    "./plugins/firebase": {
+      "import": "./dist/plugins/firebase/public.js",
+      "types": "./dist/plugins/firebase/public.d.ts"
+    },
+    "./plugins/fetch": {
+      "import": "./dist/plugins/fetch/public.js",
+      "types": "./dist/plugins/fetch/public.d.ts"
+    },
+    "./plugins/scheduler": {
+      "import": "./dist/plugins/scheduler/public.js",
+      "types": "./dist/plugins/scheduler/public.d.ts"
+    },
+    "./plugins/fs": {
+      "import": "./dist/plugins/fs/public.js",
+      "types": "./dist/plugins/fs/public.d.ts"
     }
   },
   "files": [
