+++ src/plugins/fetch/public.ts
@@ -1,0 +1,1 @@
+export { fetchPlugin } from './index.js'
