+++ src/plugins/firebase/app.ts
