+++ src/plugins/firebase/public.ts
