+++ src/plugins/firebase/types.ts
