+++ src/plugins/fs/public.ts
@@ -1,0 +1,2 @@
+export { fsPlugin } from './index.js'
+export { FileSystemPluginConfig } from './types.js'
