+++ src/plugins/scheduler/public.ts
@@ -1,0 +1,1 @@
+export { schedulerPlugin } from './index.js'
