+++ tsconfig.json
@@ -16,6 +16,7 @@
   "include": [
     "src/index.ts",
     "src/cli/index.ts",
+    "src/plugins/**/*.ts",
     "src/context/index.ts"
   ],
   "exclude": ["node_modules", "dist"]
